all in float32

+++ morphax/dmorph.py
@@ -25,7 +25,7 @@ def index_array(shape):
     a JAX numpy array
 
     """
-    return jnp.array([[x,y] for x in range(shape[0]) for y in range(shape[1])])
+    return jnp.array([[x,y] for x in range(shape[0]) for y in range(shape[1])]).astype(jnp.float32)
 
 #Transpose a structuring element
 @jax.jit
@@ -730,7 +730,7 @@ def cdmnn(type,width,size,shape_x,sample = False,p1 = 0.5,key = 0):
     max_size = max(size)
     for i in range(len(params)):
         params[i] = jnp.pad(params[i],((0,max_width - params[i].shape[0]),(0,2 - params[i].shape[1]),(0,max_size - params[i].shape[2]),(0,max_size - params[i].shape[3])),constant_values = -2)
-    params = jnp.array(params)
+    params = jnp.array(params,dtype = jnp.float32)
 
     #Numeric type
     type_num = []
@@ -775,7 +775,7 @@ def get_lim(h,key = 0):
     ch = jnp.where(h == 0,1,ch)
     ch = jnp.where((h == 2) | (h < 0),0,ch)
     ch = jnp.where((h != 0) & (h != 2) & (h >= 0),jax.random.choice(jax.random.PRNGKey(key),2,shape = (1,)),ch)
-    return ch.astype(jnp.int32)
+    return ch.astype(jnp.float32)
 
 #Visit a nighboor
 def visit_neighbor(h,params,x,y,lf):
@@ -907,12 +907,12 @@ def step_slda(params,x,y,forward,lf,type,width,size,sample = True,neighbors = 8,
                 tmp_prob = tmp_prob/jnp.sum(tmp_prob)
                 tmp_random = jax.random.choice(jax.random.PRNGKey(key[k,0]),max_size ** 2,shape = (1,),p = tmp_prob)
                 k = k + 1
-                rc = jnp.array([jnp.floor(tmp_random/max_size),tmp_random % max_size]).reshape((1,2)).astype(jnp.int32)
+                rc = jnp.array([jnp.floor(tmp_random/max_size),tmp_random % max_size]).reshape((1,2)).astype(jnp.float32)
                 #Sample limit
                 lim = get_lim(jnp.sum(par_l[node,:,rc[0,0],rc[0,1]]).reshape((1,1)),key[k,0])
                 k = k + 1
                 #Neighbor
-                nei = jnp.append(jnp.append(jnp.append(l.reshape((1,1)),node,1),lim,1),rc,1).astype(jnp.int32)
+                nei = jnp.append(jnp.append(jnp.append(l.reshape((1,1)),node,1),lim,1),rc,1).astype(jnp.float32)
                 error_tmp = visit_neighbor(nei,params,x,y,lf)
                 if hood is None:
                     hood = nei
@@ -1088,7 +1088,7 @@ def train_dmnn_slda(x,y,net,loss,xval = None,yval = None,sample = False,neighbor
                 hood = update(params,xb,yb,key[e,1])
 
                 #Update
-                hood = hood[hood[:,-1] == jnp.min(hood[:,-1]),0:-1].astype(jnp.int32)
+                hood = hood[hood[:,-1] == jnp.min(hood[:,-1]),0:-1].astype(jnp.float32)
                 if hood.shape[0] > 0:
                     hood = hood[0,:]
                     if store_jumps:
@@ -1270,7 +1270,7 @@ def train_dmnn_stack_slda(x,y,net,loss,xval = None,yval = None,sample = False,ne
                 hood = update(params,xb,yb,key[e,1])
 
                 #Update
-                hood = hood[hood[:,-1] == jnp.min(hood[:,-1]),0:-1].astype(jnp.int32)
+                hood = hood[hood[:,-1] == jnp.min(hood[:,-1]),0:-1].astype(jnp.float32)
                 if hood.shape[0] > 0:
                     hood = hood[0,:]
                     if store_jumps:
